pyproject.toml
[project]
name = "mf-auto-reg"
version = "0.1.0"
description = "A tool to automatically extract transaction information from ANA Pay email notifications in Gmail and register it in MoneyForward."
authors = [
    {name = "o3-ozono",email = "ee065195@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "google-api-python-client (>=2.163.0,<3.0.0)",
    "google-auth-httplib2 (>=0.2.0,<0.3.0)",
    "google-auth-oauthlib (>=1.2.1,<2.0.0)",
    "python-dotenv (>=1.0.1,<2.0.0)"
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = ["mf_auto_reg"]
package-dir = {"" = "src"}

[project.optional-dependencies]
dev = [
    "coverage (>=7.6.0,<8.0.0)",
]

[tool.coverage.run]
source = ["src/mf_auto_reg"]
omit = ["tests/*"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "if __name__ == .__main__.:",
    "pass",
    "raise NotImplementedError",
]

[project.scripts]
mf-auto-reg = "mf_auto_reg.gmail_to_moneyforward:main"
